split off TestTimedeltaIndex class
